move simulate_batch() to simulate.py, modify usages appropriately
