+++ examples/batched_simulation.py
@@ -10,9 +10,8 @@ from rotorpy.vehicles.crazyflie_params import quad_params as cf_quad_params
 from rotorpy.vehicles.hummingbird_params import quad_params as hb_quad_params
 from rotorpy.utils.trajgen_utils import generate_random_minsnap_traj
 from rotorpy.world import World
-from rotorpy.wind.default_winds import BatchedNoWind
-from rotorpy.batch_simulate import simulate_batch
-from rotorpy.simulate import simulate
+from rotorpy.wind.default_winds import NoWind, BatchedNoWind
+from rotorpy.simulate import simulate, simulate_batch
 from rotorpy.sensors.imu import Imu
 from rotorpy.sensors.external_mocap import MotionCapture
 from rotorpy.estimators.nullestimator import NullEstimator
